fix: corrige audio de negado não tocar
